Deleted FlyingCameraControl because there is no use for it unless we are doing a third person camera.
Instead, went with implementing the smoke emitters as controls(SmokeControl). This is implemented and works well.
My first test was attaching it to my ninja spatial. There is now a smoke trail that follows the path of the ninja.

Since this test was successful I will now attempt to add the SmokeControl in place of the existing smoke.